+++ src/index.d.ts
@@ -7,7 +7,8 @@ export { address, crypto, networks, payments, script };
 export { Block } from './block';
 /** @hidden */
 export { TaggedHashPrefix } from './crypto';
-export { Psbt, PsbtTxInput, PsbtTxOutput, Signer, SignerAsync, HDSigner, HDSignerAsync, } from './psbt';
+export { Psbt } from './psbt';
+export { PsbtTxInput, PsbtTxOutput, Signer, SignerAsync, HDSigner, HDSignerAsync, } from './psbt/interfaces';
 /** @hidden */
 export { OPS as opcodes } from './ops';
 export { Transaction } from './transaction';
